Able to read chr file.  RootNode an issue with new model format
